update logical device tests
